requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
ipython==8.12.3
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
multidict==6.0.5
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
ta==0.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
